Added pagetitle component
