+++ src/App.jsx
@@ -1,15 +1,15 @@
-import { Home } from "./components/home/Home";
 import { BrowserRouter, Routes, Route, Link } from "react-router-dom";
-import { AddCard } from "./components/addcard/AddCard";
-import "./components/style.css";
+import "./components/Buttons/style.css";
 import "./index.css";
+import { Home } from "./pages/home/Home";
+import { AddCard } from "./pages/addcard/AddCard";
 
 function App() {
   return (
     <>
       <BrowserRouter>
         <Routes>
-          <Route path="/" element={<Home />} />
+          <Route index element={<Home />} />
           <Route path="/AddCard" element={<AddCard />} />
         </Routes>
       </BrowserRouter>
