+++ src/components/Buttons/HomeButton.jsx
@@ -1,0 +1,15 @@
+import { Link } from "react-router-dom";
+
+export function HomeButton() {
+  return (
+    <>
+      <div className="button-container">
+        <Link to="/AddCard">
+          <button className="card__button add--new-card-btn">
+            ADD A NEW CARD
+          </button>
+        </Link>
+      </div>
+    </>
+  );
+}
