--- src/components/home/Home.jsx
@@ -1,13 +1,0 @@
-import { Link } from "react-router-dom";
-
-export function Home() {
-  return (
-    <div className="button-container">
-      <Link to="/AddCard">
-        <button className="card__button add--new-card-btn">
-          ADD A NEW CARD
-        </button>
-      </Link>
-    </div>
-  );
-}
