+++ src/components/pagetitle/style.css
@@ -1,0 +1,15 @@
+@import url('https://fonts.googleapis.com/css2?family=PT+Mono&family=Source+Sans+3:ital,wght@0,200..900;1,200..900&display=swap');
+
+.page-header{
+    display: flex;
+    justify-content: center;
+    height: 100px;
+    margin-top: 3rem;
+}
+
+.header-title{
+    font-family: 'Source Sans Pro', sans-serif;
+    font-weight: 700;
+    width: 13rem;
+    text-align: center;
+}
