+++ src/pages/addcard/AddCard.jsx
@@ -1,0 +1,10 @@
+import { AddCardButton } from "../../components/Buttons/AddCardButton";
+import { PageTitle } from "../../components/pagetitle/PageTitle";
+export function AddCard() {
+  return (
+    <>
+      <PageTitle title={"ADD A NEW BANK CARD"} />
+      <AddCardButton />
+    </>
+  );
+}
