+++ src/pages/home/Home.jsx
@@ -1,0 +1,11 @@
+import { HomeButton } from "../../components/Buttons/HomeButton";
+import { PageTitle } from "../../components/pagetitle/PageTitle";
+
+export function Home() {
+  return (
+    <>
+      <PageTitle title={"E-WALLET"} />
+      <HomeButton />
+    </>
+  );
+}
